catch infinite loop in summary_qa.create_splits()
There was an edge case infinite loop before, when there is a sentence >= 512 tokens. There is no way to split this sentence further, so it loops forever.

This `elif` is a quick fix -- it will just not process this text, returning `[]`. The assumption is that such a long sentence is probably "unnatural" text, so we don't use it.